python-consumer/requirements.txt
kafka-python
sentence-transformers
qdrant-client
python-dotenv